Add new Playerbot command: Neutralize
* Add new command to playerbot: neutralize. This command will determine
if the target is eligible to be put out of combat by using a dedicated
Neutralize() function from each bot class
Note that neutralizing a target will put bot in combat with that
target. Thus, if they is no other target in the bot enemy list, the bot
will attack the target right after neutralizing it.